# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  v1.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: string
      error:
        type: string
    type: object
...
info:
  contact:
    name: OWASP Amass Project
    url: https://github.com/owasp-amass/amass
  description: |-
    HTTP API for managing Amass Engine sessions and submitting Open Asset Model (OAM) assets.
    Create a session with an engine config, query session stats/scope, and ingest typed OAM assets (single or bulk).
    Includes a WebSocket endpoint for streaming session logs.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  termsOfService: https://owasp.org/www-project-amass/
  title: Amass Engine API (v1)
  version: "1.0"
paths:
...
  /v1/sessions/{session_token}/ws/logs:
    get:
      description: |-
        Upgrades the HTTP connection to a WebSocket and streams session log lines as UTF-8 text frames.
        The server sends periodic ping frames (~30s) and expects pong responses; idle connections may be closed.
      parameters:
      - description: Session token (UUID)
        in: path
        name: session_token
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade)
          headers:
            Connection:
              description: Upgrade
              type: string
            Upgrade:
              description: websocket
              type: string
        "400":
          description: Invalid session token
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Stream session logs (WebSocket)
      tags:
      - sessions
swagger: "2.0"
...